backend/requirements.txt
alembic==1.14.0
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssbeautifier==1.15.1
djlint==1.36.1
dnspython==2.7.0
EditorConfig==0.12.4
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flower==2.0.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.10.0
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
packaging==24.2
pathspec==0.12.1
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
six==1.16.0
SQLAlchemy==2.0.36
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
